# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hawk-Eye Control Plane",
    "description": "Insider & privileged-user fraud detection — FastAPI control plane (ALERT-ONLY).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/entities/{entity_id}": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Get Entity",
        "operationId": "get_entity_api_v1_entities__entity_id__get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entities/{entity_id}/timeline": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Get Timeline",
        "operationId": "get_timeline_api_v1_entities__entity_id__timeline_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityTimeline"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entities/{entity_id}/graph": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Get Graph",
        "operationId": "get_graph_api_v1_entities__entity_id__graph_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityGraph"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entities/{entity_id}/peers": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Get Peers",
        "operationId": "get_peers_api_v1_entities__entity_id__peers_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PeerComparison"
                  },
                  "title": "Response Get Peers Api V1 Entities  Entity Id  Peers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entities/{entity_id}/unmask": {
      "post": {
        "tags": [
          "entities"
        ],
        "summary": "Unmask",
        "description": "Re-identify tokenized PII (audited). The RM must provide a justification (case-scoped).",
        "operationId": "unmask_api_v1_entities__entity_id__unmask_post",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnmaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnmaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Models",
        "operationId": "list_models_api_v1_models_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ModelInfo"
                  },
                  "title": "Response List Models Api V1 Models Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{model_id}/promote": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Promote Model",
        "operationId": "promote_model_api_v1_models__model_id__promote_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Artifact version to promote",
              "title": "Version"
            },
            "description": "Artifact version to promote"
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoteResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/drift": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Drift",
        "operationId": "drift_api_v1_drift_get",
        "parameters": [
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "l3_lightgbm",
              "title": "Model Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriftReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DriftReport": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "data_drift_psi": {
            "type": "number",
            "title": "Data Drift Psi"
          },
          "concept_drift": {
            "type": "number",
            "title": "Concept Drift"
          },
          "drift_crossed": {
            "type": "boolean",
            "title": "Drift Crossed"
          },
          "window": {
            "type": "string",
            "title": "Window",
            "default": "rolling_30d"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "version",
          "data_drift_psi",
          "concept_drift",
          "drift_crossed"
        ],
        "title": "DriftReport"
      },
      "EntityGraph": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "ring_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ring Id"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/GraphNode"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/GraphEdge"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "entity_id"
        ],
        "title": "EntityGraph"
      },
      "EntityProfile": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id",
            "examples": [
              "EMP-7f3a"
            ]
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "dept": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dept"
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "peer_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peer Group"
          },
          "tenure_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenure Days"
          },
          "privileged_flag": {
            "type": "boolean",
            "title": "Privileged Flag",
            "default": false
          },
          "leaver_flag": {
            "type": "boolean",
            "title": "Leaver Flag",
            "default": false
          },
          "risk_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "default": 0
          },
          "severity": {
            "$ref": "#/components/schemas/Severity",
            "default": "low"
          },
          "open_alerts": {
            "type": "integer",
            "title": "Open Alerts",
            "default": 0
          },
          "pii_tokenized": {
            "type": "boolean",
            "title": "Pii Tokenized",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "entity_id"
        ],
        "title": "EntityProfile"
      },
      "EntityTimeline": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "entity_id"
        ],
        "title": "EntityTimeline"
      },
...
      "GraphNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "employee | beneficiary | account | device | ip"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "risk": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind"
        ],
        "title": "GraphNode"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelInfo": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id",
            "examples": [
              "l3_gbdt"
            ]
          },
          "layer": {
            "type": "string",
            "title": "Layer",
            "examples": [
              "L3_gbdt"
            ]
          },
          "version": {
            "type": "string",
            "title": "Version",
            "examples": [
              "2026.06.30-1"
            ]
          },
          "stage": {
            "type": "string",
            "title": "Stage",
            "description": "Production | Staging | Challenger | Archived"
          },
          "signed": {
            "type": "boolean",
            "title": "Signed",
            "description": "Signature verified against registry (BACKEND-16)"
          },
          "training_data_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Data Hash"
          },
          "feature_set_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Set Version"
          },
          "approving_reviewer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approving Reviewer"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "layer",
          "version",
          "stage",
          "signed"
        ],
        "title": "ModelInfo"
      },
...
      "PeerComparison": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "peer_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peer Group"
          },
          "dimension": {
            "type": "string",
            "title": "Dimension",
            "description": "The flagged dimension being compared"
          },
          "actor_value": {
            "type": "number",
            "title": "Actor Value"
          },
          "peer_mean": {
            "type": "number",
            "title": "Peer Mean"
          },
          "peer_p95": {
            "type": "number",
            "title": "Peer P95"
          },
          "z_score": {
            "type": "number",
            "title": "Z Score"
          },
          "is_outlier": {
            "type": "boolean",
            "title": "Is Outlier"
          }
        },
        "type": "object",
        "required": [
          "entity_id",
          "dimension",
          "actor_value",
          "peer_mean",
          "peer_p95",
          "z_score",
          "is_outlier"
        ],
        "title": "PeerComparison"
      },
      "PromoteRequest": {
        "properties": {
          "to_stage": {
            "type": "string",
            "title": "To Stage",
            "default": "Production",
            "examples": [
              "Production",
              "Staging"
            ]
          },
          "signoff_by": {
            "type": "string",
            "title": "Signoff By",
            "description": "Second-person approver (≠ requester) — SoD"
          },
          "canary_percent": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Canary Percent",
            "default": 10
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "signoff_by"
        ],
        "title": "PromoteRequest",
        "description": "Promotion requires a sign-off by a second person (SoD; promoter ≠ approver)."
      },
      "PromoteResult": {
        "properties": {
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "canary_percent": {
            "type": "integer",
            "title": "Canary Percent"
          },
          "signature_verified": {
            "type": "boolean",
            "title": "Signature Verified"
          },
          "audit_id": {
            "type": "string",
            "title": "Audit Id"
          }
        },
        "type": "object",
        "required": [
          "model_id",
          "version",
          "stage",
          "canary_percent",
          "signature_verified",
          "audit_id"
        ],
        "title": "PromoteResult"
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "Severity"
      },
...
      "TimelineEvent": {
        "properties": {
          "ts": {
            "type": "string",
            "title": "Ts"
          },
          "lane": {
            "type": "string",
            "title": "Lane",
            "description": "transaction | access | data | change | hr"
          },
          "verb": {
            "type": "string",
            "title": "Verb"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "amount_inr": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Inr"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "lane",
          "verb"
        ],
        "title": "TimelineEvent",
        "description": "One row on the unified transaction+access+data+change timeline."
      },
...
      "UnmaskRequest": {
        "properties": {
          "tokens": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tokens",
            "description": "Tokens to re-identify; empty = all on the entity"
          },
          "justification": {
            "type": "string",
            "title": "Justification",
            "description": "Case-scoped reason (required for Relationship Manager, logged)",
            "default": ""
          }
        },
        "type": "object",
        "title": "UnmaskRequest"
      },
      "UnmaskResponse": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "mapping": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Mapping",
            "description": "token → real value (audited; never logged)"
          },
          "audit_id": {
            "type": "string",
            "title": "Audit Id"
          }
        },
        "type": "object",
        "required": [
          "entity_id",
          "audit_id"
        ],
        "title": "UnmaskResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}